武器の合成と売却、初期武器選択など

+++ Brotatelike/Assets/Scripts/HealthComponent.cs
@@ -74,7 +74,7 @@ public class HealthComponent : MonoBehaviour, IDamageable
 
     private void RemoveHealth(float amount)
     {
-        currentHealth -= amount;
+        currentHealth -= (int)amount;
 
         if (IsDead)
         {
